+++ tools/src/Prepare.hx
@@ -8,7 +8,7 @@ import haxe.xml.Fast;
 class Prepare {
 
 	public function new(xml:Fast, app:String, debug:Bool) {
-		sys.FileSystem.deleteFile('libcache.js');
+		if (sys.FileSystem.exists('libcache.js')) sys.FileSystem.deleteFile('libcache.js');
 		Sys.println('update haxelib');
 		for (node in xml.node.haxelib.nodes.lib) {
 			var args = ['install'];
@@ -16,7 +16,8 @@ class Prepare {
 			args.push('--always');
 			Sys.command('haxelib', args);
 		}
-		new Build(xml, app, debug).writeConfigIfNeed();
+		if (app != null)
+			new Build(xml, app, debug).writeConfigIfNeed();
 	}
 	
 }
